Remove admin, core package
- all the operations in admin subpackage is now top level on Client
- all the operations in core subpackage is now in core.go
- core package is still there because tests are not migrated yet
