+++ solr/fixture/facet.json
