+++ solr/query_test.go
@@ -1,4 +1,4 @@
-package search
+package solr
 
 import "testing"
